requirements.txt
bottle
pillow
numpy
mysql-connector-python